requirements.txt
bs4
faker
pyyaml
colorama
tls_client
anticaptchaofficial